Extract common code to new resource class.
